重写手册的过程中，对代码也做了些调整，如下： 1、修改了大量注释，不一一说明了。 2、删除了ring.c中一处重复函数，Ring_Capacity2函数文不对题，改名为Ring_CheckFree 3、handle_IsBlockComplete函数从device.c移到objhandle.c，并改名为 test_IsBlockComplete 4、device.c中：    __devcntl函数中，删除了F_SETTIMEOUT、F_GETTIMEOUT两个命令，转移到fcntl函数中实现了。    __devcntl函数不再支持 ioctl函数。    device.c的dev_GetDrvTagFromObj函数被删除，统一使用dev_GetDrvTag函数    device.c中改了几个函数名。 5、objhandle.c中：    添加函数：test_IsBlockComplete    handle_setmultiplex改名为 __handle_SetMultiplexHead，handle_GetMultiplexHead改名为 __handle_GetMultiplexHead，均不再作为API    fcntl函数直接实现了F_SETTIMEOUT和F_GETTIMEOUT命令，不再需要下级文件系统实现。    修改了fcntl和ioctl函数实现过程。 6、stdio.c中删除了无用的 __stdio_timeout 函数，并取消了 F_SETTIMEOUT和F_GETTIMEOUT 实现，在fcntl中已经实现。 7、修改了stdio模块的配置方法。 8、修改了timer.c的配置方式，并消掉其一个bug。 9、djyos.c中：    Djy_EventExit函数改为__Djy_EventExit，不再是API    Djy_EventSessionComplete改名为 Djy_EventComplete    Djy_EventComplete改名为__Djy_EventFinal，不再是API 10、object.c中，增加函数 obj_rename

+++ djyos/time/systime.c
@@ -64,7 +64,7 @@
 #include "systime.h"
 #include "component_config_time.h"
 
-extern u64 __DjyGetSysTime(void);
+extern s64 __DjyGetSysTime(void);
 #if (!CN_USE_TICKLESS_MODE)
 extern s64  g_s64OsTicks;               //操作系统运行ticks数
 #endif
@@ -76,6 +76,12 @@ static u32 s_u32SysTimeCycle = CN_CFG_TICK_US;  //系统时钟走时周期，CN_CFG_TICK_
                                                 //是选tick时基默认值
 static s64 s_s64sysTimeMajor = 0;       //时钟高位
 static u32 s_u32BakCounter = 0;         //计数值的备份
+
+//__attribute__((weak)) s64 __DjyGetSysTime(void)
+//{
+//    debug_printf("systime", "请bsp提供 __DjyGetSysTime函数")
+//    return 0;
+//}
 
 // =============================================================================
 // 函数功能：连接专用定时器到系统时间
@@ -90,7 +96,7 @@ static u32 s_u32BakCounter = 0;         //计数值的备份
 void SysTimeConnect(fntSysTimeHard32 GetSysTime32,fntSysTimeHard64 GetSysTime64,
                     u32 Freq,u32 Cycle)
 {
-    //不要判SysTime是否为NULL，NULL也是一个正确选项。
+    //不要判GetSysTime32和GetSysTime64是否为NULL，NULL也是一个正确选项。
     fnSysTimeHard32 = GetSysTime32;
     fnSysTimeHard64 = GetSysTime64;
     s_u32SysTimeFreq = Freq;
@@ -110,7 +116,7 @@ s64 DjyGetSysTime(void)
     atom_low_t atom;
 
     if((fnSysTimeHard32 == NULL) && (fnSysTimeHard64 == NULL))
-        return (s64)__DjyGetSysTime();
+        return __DjyGetSysTime();
     else if(fnSysTimeHard32 != NULL)
     {
         atom = Int_LowAtomStart();
@@ -148,7 +154,7 @@ s64 DjyGetSysTimeCycle(void)
     atom_low_t atom;
 
     if((fnSysTimeHard32 == NULL) && (fnSysTimeHard64 == NULL))
-        return (s64)__DjyGetSysTime();
+        return __DjyGetSysTime();
     else if(fnSysTimeHard32 != NULL)
     {
         atom = Int_LowAtomStart();
